Camera updated!
Sadly, I couldn't get the multiple bounds system working like the draft; the bounds were
too finicky and oddly enough the camera was actually getting stuck to them instead of
going around them.

So, I created the Camera Waypoint system to tackle this problem.

Camera Waypoints are GameObjects which, upon the player entering their trigger collider,
send a message to the camera to change its behavior.

The idea is that these can be placed around doorways to other areas to make the camera's
bounds change depending upon the room the player enters.

-Files Edited:
	-CameraBehavior.cs
		-Min/Max bounds system overhauled and expanded; the system now uses camera
		 "waypoints"
	-CameraWaypoint.cs
		-created to be used with the new camera positioning system.
	-CameraWaypoint
		-created

Signed-off-by: Devin Aldrich <19yoshi95@gmail.com>